Library: typecast in comparison of int and long u int

+++ YACReaderLibrary/yacreader_local_server.cpp
@@ -104,7 +104,7 @@ void YACReaderClientConnectionWorker::run()
 	int dataAvailable = 0;
 	QByteArray packageSize;
     clientConnection->waitForReadyRead(1000);
-	while(packageSize.size() < sizeof(quint32) && tries < 20)
+	while(((long unsigned int)packageSize.size() < sizeof(quint32)) && (tries < 20))
 	{
 		packageSize.append(clientConnection->read(sizeof(quint32) - packageSize.size()));
 		clientConnection->waitForReadyRead(100);
